Zoom map implemented and Refactored Geo Scroll
