+++ src/pages/GeoIntro.tsx
@@ -4,7 +4,7 @@ import '@mantine/core/styles.css';
 import NavbarPlain from '@/components/NavbarPlain';
 import HeroLayout from '@/components/HeroLayout';
 import { animateScroll as scroll } from 'react-scroll';
-import GeoScrollytelling from '@/components/GeoScrollytelling';
+import GeoScrollytelling from '@/pages/geoScrollytelling/GeoScrollytelling';
 
 const GeoIntro = () => {
   useEffect(() => {
